Fixed a few CDP1802 opcodes.
There are a few errors in the numeric order table of opcodes in the
back of the 1976 user manual for the CDP1802, part number MPM-201A.
REQ and SEQ are swapped, and ADCI was listed as ADDI. Assuming
that they are correct in earlier parts of the book, and that agrees
with what the asmx assembler expects. I have not personally verified
the correct REQ/SEQ opcodes on a real CDP1802 processor yet.